V0.0.5 - Dockerizing Continue
