+++ .dockerignore
@@ -1,0 +1,5 @@
+Dockerfile
+.dockerignore
+node_modules
+npm-debug.log
+dist
