--- .env.development
@@ -1,10 +1,0 @@
-DB_HOST=localhost
-DB_PORT=5432
-DB_USERNAME=postgres
-DB_PASSWORD=postgres
-DB_NAME=blog_test
-DB_CHARSET=utf8mb4
-DB_TIMEZONE=+09:00
-
-
-GRPC_CONNECTION_URL=localhost:3002
