+++ .gitignore
@@ -1,7 +1,7 @@
 # compiled output
 /dist
 /node_modules
-/db/migrations
+/migrations
 
 # Logs
 logs
@@ -17,6 +17,7 @@ lerna-debug.log*
 
 # files
 /uploads
+/pgdata
 
 # Tests
 /coverage
