+++ docker-compose.yml
@@ -1,0 +1,92 @@
+# version: '3.7'
+
+# services:
+#   db:
+#     image: postgres
+#     restart: always
+#     environment:
+#       - POSTGRES_HOST=postgres
+#       - POSTGRES_PORT=5432
+#       - POSTGRES_DB= blog_test
+#       - POSTGRES_USER= postgres
+#       - POSTGRES_PASSWORD=postgres
+#     container_name: postgres
+#     volumes:
+#       - ./pgdata:/var/lib/postgresql/data
+#     ports:
+#       - '5432:5432'
+
+#   app:
+#     build:
+#       context: .
+#       dockerfile: Dockerfile
+#     container_name: nest-docker-postgres
+#     environment:
+#       - PORT=3000
+#     ports:
+#       - '8080:3000'
+#     depends_on:
+#       - db
+#     volumes:
+#       - ./src:/app/src
+
+#   pgadmin:
+#     image: dpage/pgadmin4
+#     restart: always
+#     container_name: nest-pgadmin4
+#     environment:
+#       - PGADMIN_DEFAULT_EMAIL=admin@admin.com
+#       - PGADMIN_DEFAULT_PASSWORD=pgadmin4
+#     ports:
+#       - '5050:80'
+#     depends_on:
+#       - db
+
+
+version: '3.7'
+
+services:
+  db:
+    container_name: postgres
+    image: postgres
+    restart: always
+    ports:
+      - '5432:5432'
+    environment:
+      POSTGRES_DB: "blog_test"
+      POSTGRES_PASSWORD: "postgres"
+      POSTGRES_USER: "postgres"
+    volumes:
+      - ./pgdata:/var/lib/postgresql/data
+
+  app:
+    build:
+      context: .
+      dockerfile: Dockerfile
+    container_name: nest-docker-postgres
+    environment:
+      DB_HOST: localhost
+      DB_PORT: 5432
+      DB_NAME: blog_test
+      DB_USERNAME: postgres
+      DB_PASSWORD: postgres
+      PORT: 3000
+      DATABASE_URL: "postgres://postgres:postgres@db:5432/blog_test"
+    ports:
+      - '8080:3000'
+    depends_on:
+      - db
+    volumes:
+      - ./src:/app/src
+
+  pgadmin:
+    image: dpage/pgadmin4
+    restart: always
+    container_name: nest-pgadmin4
+    environment:
+      - PGADMIN_DEFAULT_EMAIL=admin@admin.com
+      - PGADMIN_DEFAULT_PASSWORD=pgadmin4
+    ports:
+      - '5050:80'
+    depends_on:
+      - db
