+++ dockerfile
@@ -1,0 +1,13 @@
+FROM node:18.18.1
+
+WORKDIR /app
+
+COPY package*.json ./
+
+RUN npm install
+
+COPY . .
+
+RUN npm run build
+
+CMD [ "npm", "run", "start:dev" ]
