+++ src/main.ts
@@ -2,10 +2,10 @@ import { NestFactory } from '@nestjs/core';
 import { DocumentBuilder, SwaggerModule } from '@nestjs/swagger';
 import { ValidationPipe } from '@nestjs/common';
 import { AppModule } from './app.module';
-
 import * as morgan from 'morgan';
 
 async function bootstrap() {
+  console.log('main');
   const app = await NestFactory.create(AppModule);
 
   const prefix = '/api/v1';
